feat(istatistik): teslimat grubu bazlı tahmin vs gerçekleşen analizi + şarj düzeltmesi
- Yeni İstatistik sekmesi: tool tahmini (şarj sonu, yoksa sim) vs gerçekleşen palet
- Gruplama güvenilmez teslimat adı yerine kalıcı "Teslimat Grubu" (statFolder) ile
- Geçmiş modalında gerçekleşen palet + zorunlu grup girişi; ondalık destekli; "Kaydedildi" toast
- Genel KPI + yatay kaydırılabilir grup-özet grafiği (gruba tıkla→detay) + sıralanabilir/satır-içi düzenlenebilir tablo + dönem filtresi
- Modalda şarj sonu tahmini palet + şarjlı/şarjsız rozeti
- FIX: aktif şarj Geçmiş'ten açılınca TES id'si şarjın deliveryId'siyle eşleşmediğinden görünmüyordu → isim+gün ile bulunuyor (veri bütünlüğü)
- Mevcut palet ve gerçekleşen palet ondalık girilebilir

+++ src/sim.js
@@ -303,7 +303,8 @@ export async function runSim() {
   if (!rawLines.length) { showWarn(t('warn_no_input')); return; }
   if (configStore.PALET_VOL <= 0) { showWarn(t('warn_no_vol')); return; }
 
-  const existingPalet = Math.max(0, parseInt(document.getElementById('existingPaletInput').value) || 0);
+  // Mevcut palet tam sayı olmak zorunda değil — saha durumlarında ondalık girilebilir.
+  const existingPalet = Math.max(0, parseFloat(document.getElementById('existingPaletInput').value) || 0);
 
   uiStore.simRunning = true;
   const btn = document.getElementById('runBtn');
